requirements.txt
aiogram==3.19.0
openai==1.72.0
python-dotenv==1.1.0
flask==2.3.3
aiohttp==3.11.16
aiofiles==24.1.0
